Rename rotations -> orientations for consistency
