+++ tests/integration/conftest.py
@@ -8,14 +8,14 @@ from gemdat.io import load_known_material
 from gemdat.jumps import Jumps
 from gemdat.path import find_best_perc_path, free_energy_graph
 from gemdat.rdf import radial_distribution
-from gemdat.rotations import Orientations
+from gemdat.orientations import Orientations
 from gemdat.shape import ShapeAnalyzer
 from gemdat.trajectory import Trajectory
 from gemdat.volume import trajectory_to_volume
 
 DATA_DIR = Path(__file__).parents[1] / 'data'
 VASP_XML = DATA_DIR / 'short_simulation' / 'vasprun.xml'
-VASP_ROTO_CACHE = DATA_DIR / 'short_simulation' / 'vasprun_rotations.cache'
+VASP_ORI_CACHE = DATA_DIR / 'short_simulation' / 'vasprun_rotations.cache'
 
 
 def pytest_configure():
@@ -25,8 +25,8 @@ def pytest_configure():
         ('Simulation data from vasprun.xml example is required for this test. '
          'Run `git submodule init`/`update`, and extract using `tar -C tests/data/short_simulation '
          '-xjf tests/data/short_simulation/vasprun.xml.bz2`'))
-    pytest.vasprotocache_available = pytest.mark.skipif(
-        not VASP_ROTO_CACHE.exists(),
+    pytest.vasporicache_available = pytest.mark.skipif(
+        not VASP_ORI_CACHE.exists(),
         reason=
         ('Simulation data from vasprun_rotations.cache example is required for this test. '
          'Run `git submodule init`/`update`, and extract using `tar -C tests/data/short_simulation '
@@ -41,8 +41,8 @@ def vasp_traj():
 
 
 @pytest.fixture(scope='module')
-def vasp_traj_rotations():
-    trajectory = Trajectory.from_cache(VASP_ROTO_CACHE)
+def vasp_traj_orientations():
+    trajectory = Trajectory.from_cache(VASP_ORI_CACHE)
     return trajectory
 
 
@@ -135,11 +135,11 @@ def vasp_F_graph(vasp_path_vol):
 
 
 @pytest.fixture(scope='module')
-def vasp_orientations(vasp_traj_rotations):
+def vasp_orientations(vasp_traj_orientations):
     central_atoms = 'S'
     satellite_atoms = 'O'
     n_expected_neigh = 8
-    orientations = Orientations(vasp_traj_rotations, central_atoms,
+    orientations = Orientations(vasp_traj_orientations, central_atoms,
                                 satellite_atoms, n_expected_neigh)
 
     return orientations
