+++ tests/integration/rotations_test.py
@@ -5,14 +5,14 @@ from math import isclose
 import numpy as np
 import pytest
 
-from gemdat.rotations import calculate_spherical_areas
+from gemdat.orientations import calculate_spherical_areas
 
 matrix = np.array(
     [[1 / 2**0.5, -1 / 6**0.5, 1 / 3**0.5],
      [1 / 2**0.5, 1 / 6**0.5, -1 / 3**0.5], [0, 2 / 6**0.5, 1 / 3**0.5]], )
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_normalize(vasp_orientations):
     ret = vasp_orientations.normalize()
     assert isclose(ret.vectors.std(), 0.5773501929401034)
@@ -21,7 +21,7 @@ def test_normalize(vasp_orientations):
                                atol=1e-06)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_conventional(vasp_orientations):
     ret = vasp_orientations.transform(matrix=matrix)
     assert isclose(ret.vectors.std(), 0.8668510720176622)
@@ -30,7 +30,7 @@ def test_conventional(vasp_orientations):
                                atol=1e-06)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_symmetrize(vasp_orientations):
     ret = vasp_orientations.symmetrize(sym_group='m-3m')
     assert isclose(ret.vectors.std(), 0.8668511628167984)
@@ -39,7 +39,7 @@ def test_symmetrize(vasp_orientations):
                                atol=1e-06)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_normalize_symmetrize(vasp_orientations):
     ret = vasp_orientations.normalize().symmetrize(sym_group='m-3m')
     assert isclose(ret.vectors.std(), 0.577350269189626)
@@ -48,7 +48,7 @@ def test_normalize_symmetrize(vasp_orientations):
                                atol=1e-06)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_normalize_conventional_symmetrize(vasp_orientations):
     ret = vasp_orientations.normalize().transform(matrix=matrix).symmetrize(
         sym_group='m-3m')
@@ -58,7 +58,7 @@ def test_normalize_conventional_symmetrize(vasp_orientations):
                                atol=1e-06)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_conventional_normalize_symmetrize(vasp_orientations):
     ret = vasp_orientations.transform(matrix=matrix).normalize().symmetrize(
         sym_group='m-3m')
@@ -68,7 +68,7 @@ def test_conventional_normalize_symmetrize(vasp_orientations):
                                atol=1e-06)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_fractional_coordinates(vasp_orientations):
     direction = vasp_orientations._fractional_directions(
         vasp_orientations._distances)
@@ -77,7 +77,7 @@ def test_fractional_coordinates(vasp_orientations):
     assert isclose(direction[3, 12, 2], -0.10902666)
 
 
-@pytest.vasprotocache_available  # type: ignore
+@pytest.vasporicache_available  # type: ignore
 def test_matching_matrix(vasp_orientations):
     distance = np.array([[1.0, 2.0, 1.5, 0.5, 3.0], [2.0, 1.0, 0.5, 1.5, 3.0]])
     frac_coord_cent = np.array([[[0.3, 0.11, 0.78], [1.6, 1.0, 2.3]]])
